+++ src/Source.h
@@ -47,6 +47,37 @@ public:
     return cur;
   }
 
+  static const char* skipBom(const char* begin, const char* end) {
+
+    /* Unicode Byte Order Marks
+       https://en.wikipedia.org/wiki/Byte_order_mark#Representations_of_byte_order_marks_by_encoding
+
+       EF BB BF:    UTF-8
+       FF FE 00 00: UTF-32LE
+       FF FE:       UTF-16LE
+       FE FF:       UTF-16BE
+       00 00 FE FF: UTF-32BE
+   */
+    boost::iterator_range<const char*> haystack(begin, end);
+
+    if (boost::starts_with(haystack, "\xEF\xBB\xBF")) {
+      return begin + 3;
+    }
+    if (boost::starts_with(haystack, "\xFF\xFE\x00\x00")) {
+      return begin + 4;
+    }
+    if (boost::starts_with(haystack, "\xFF\xFE")) {
+      return begin + 2;
+    }
+    if (boost::starts_with(haystack, "\xFE\xFF")) {
+      return begin + 2;
+    }
+    if (boost::starts_with(haystack, "\x00\x00\xFE\xFF")) {
+      return begin + 4;
+    }
+    return begin;
+  }
+
   static SourcePtr create(Rcpp::List spec);
 
 private:
@@ -59,5 +90,3 @@ private:
 };
 
 #endif
-
-