+++ NEWS.md
@@ -1,9 +1,11 @@
 # readr 0.2.2.9000
+* Skip Unicode byte order markers if they exist (#263, @jimhester).
 
 * Parse doubles with `boost::spirit::qi::long_double` to work around a bug in the
   spirit library when parsing large numbers (#412, @jimhester).
 
 * Supports reading into long vectors (#309, @jimhester).
+
 * `default_locale()` now sets the default locale in `readr.default_locale`
   rather than regenerating it for each call. (#416, @jimhester).